Move game configuration to conf.lua
- This should cause the console to have useful data on Windows
