+++ src/constants/screen_constants.lua
@@ -1,0 +1,8 @@
+-- Constants for the screen. Kept in a separate file
+-- so that conf.lua can access them too
+
+VIRTUAL_WIDTH = 768
+VIRTUAL_HEIGHT = 432
+
+WINDOW_WIDTH = 1280
+WINDOW_HEIGHT = 720
